requirements.txt
python-telegram-bot>=13.4
redis
requests